+++ controllers/utilities/index.js
@@ -1,7 +1,7 @@
-module.exports = exports = { _$: {
+module.exports = exports = { _U_: {
     schema  : require('./schema'),
     string  : require('./string'),
     object  : require('./object'),
     error   : require('./error'),
 }};
-exports._test = { ...exports._$.schema, ...exports._$.string, ...exports._$.object, ...exports.error };
+exports._test = { ...exports._U_.schema, ...exports._U_.string, ...exports._U_.object, ...exports.error };