+++ controllers/views/handler.js
@@ -1,5 +1,5 @@
-const { _$ } = require('../modules/');
-const { _md } = require('../middleware/plugins');
+const { _U_ } = require('../utilities/');
+const { _M_ } = require('../middleware/plugins');
 module.exports = exports = {};
 
 
@@ -14,14 +14,14 @@ exports.postHandler = (template, post, meta) => async (req, res, next) => {
     const $meta = await meta ? meta : req.session.view ? req.session.view.meta : {};
     delete req.session.view;
 
-    if ($post && $post.title !== undefined) _md.setTitleTag($post.title, { root: true })(req, res);
+    if ($post && $post.title !== undefined) _M_.setTitleTag($post.title, { root: true })(req, res);
     if ($meta) {
         const urlBase = res.baseUrl + `?num=' + ${$meta.num} + '&page='`;
         if ($meta.now > 1)          res.locals._view.prev = urlBase + ($meta.now - 1);
         if ($meta.now < $meta.end)  res.locals._view.next = urlBase + ($meta.now + 1);
     }
 
-    if ($post === null) throw new _$.error.HttpError(404);
+    if ($post === null) throw new _U_.error.HttpError(404);
     else return res.render($template, { meta: $meta, post: $post });
 };
 
@@ -35,7 +35,7 @@ const terminal = {};
 // gateway
 exports.errorHandler = (err, req, res, next) => {     // todo: error handler separations
     if (['dev', 'test'].indexOf(process.env.NODE_ENV) !== -1) console.log(err.stack);
-    if (_$.error.hasOwnProperty(err.name) && terminal[err.name]) {
+    if (_U_.error.hasOwnProperty(err.name) && terminal[err.name]) {
         return terminal[err.name](err, req, res, next);
     } else return res.render('./theme/error', { err });
 };
@@ -65,15 +65,15 @@ terminal.MongoError = (err, req, res, next) => {
 };
 
 terminal.HttpError = (err, req, res, next) => {
-    return _md.doNotCrawled(req, res, () => {
+    return _M_.doNotCrawled(req, res, () => {
         return res.status(err.status).render('./theme/error', {err});
     });
 };
 
 terminal.TemplateError = (err, req, res, next) => {
-    return _md.doNotCrawled(req, res, () => {
+    return _M_.doNotCrawled(req, res, () => {
         // todo: log the message and call the admin
         // todo: guidance for the client
-        return res.status(500).send(`<h1>${new _$.error.HttpError(500).message}</h1>`);
+        return res.status(500).send(`<h1>${new _U_.error.HttpError(500).message}</h1>`);
     });
 };